fix: test manual prerender search
